Introduce MetalLB
MetalLB will be in charge of assigning external IP to services within
the cluster.

However, I have a small issue with port forwarding in the Google WiFi.
It does not allow routing traffic to virtual non-device IP addresses,
so currently I still use `hostPort` for external HTTP traffic.
Hopefully this will be fixed sooner or later.

Closes #63